UserRepository

+++ CleanArchDemo.Applications/Command/Users/AddUserCommand.cs
@@ -12,10 +12,10 @@ namespace CleanArchDemo.Applications.Command.Users
 {
     public class AddUserCommand(Id id,Name name,Age age,Email email):IRequest<UserDto>
     {
-        public double Id { get; set; } = id;
+        public int Id { get; set; } = id;
         public string Name { get; set; } = name;
         public string Email { get; set; } = email;
-        public Double Age { get; set; } = age;
+        public int Age { get; set; } = age;
 
     }
 }
