+++ CleanArchDemo.Applications/Command/Users/UpdateUserCommand.cs
@@ -6,10 +6,10 @@ namespace CleanArchDemo.Applications.Command.Users
 {
     public class UpdateUserCommand(Id id, Name name, Age age, Email email) : IRequest<UserDto>
     {
-        public double id { get; set; } = id;
+        public int id { get; set; } = id;
         public string name { get; set; } = name;
         public string Email { get; set; } = email;
-        public Double Age { get; set; } = age
+        public int Age { get; set; } = age
 
     }
 }
